+++ src/app/(main)/clubs/_components/ClubsHeader.tsx
@@ -1,15 +1,15 @@
 "use client";
 
-import Image from "next/image";
+import SmallButton from "@/commons/Button/SmallButton";
+import HeaderSection from "@/commons/Header/HeaderSection";
+import SkeletonHeader from "@/commons/Header/SkeletonHeader";
+import { useUserProfile } from "@/hooks/useUserProfile";
+import { useUserStore } from "@/store";
 import { createClient } from "@/utils/supabase/client";
+import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
-import SmallButton from "../Button/SmallButton";
 import EditNickname from "./EditNickname";
-import HeaderSection from "./HeaderSection";
-import { useUserStore } from "@/store";
-import { useUserProfile } from "@/hooks/useUserProfile";
-import SkeletonHeader from "./SkeletonHeader";
 
 const ClubsHeader = () => {
   useUserProfile();
